# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Assistant",
    "description": "Backend API for AI-powered document processing, semantic search, document management, and chat.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/documents/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get user documents",
        "description": "Return all documents uploaded by the authenticated user.",
        "operationId": "get_documents_api_documents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Documents Api Documents  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete document",
        "description": "Delete an authenticated user's document, its stored vectors, and its physical file.",
        "operationId": "delete_document_api_documents__document_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Document Api Documents  Document Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/": {
      "post": {
        "tags": [
          "AI Chat"
        ],
        "summary": "Chat with AI",
        "description": "Answer questions using uploaded documents, Groq directly, or smart retrieval with Groq fallback.",
        "operationId": "chat_api_chat__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Question",
            "description": "User question."
          },
          "mode": {
            "type": "string",
            "enum": [
              "document",
              "groq",
              "smart"
            ],
            "title": "Mode",
            "description": "Chat processing mode.",
            "default": "document"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "Optional document ID used to restrict chat to a specific uploaded document."
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest",
        "description": "Request payload for AI chat.\n\nModes:\n- document: Answer using uploaded document context.\n- groq: Answer directly using the configured Groq LLM.\n- smart: Use document context when relevant and Groq otherwise."
      },
      "ChatResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "title": "Answer"
          },
          "mode": {
            "type": "string",
            "enum": [
              "document",
              "groq",
              "smart"
            ],
            "title": "Mode"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "answer",
          "mode"
        ],
        "title": "ChatResponse",
        "description": "Response returned by the AI chat endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceDocument": {
        "properties": {
          "document_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename"
        ],
        "title": "SourceDocument",
        "description": "Document source used to generate an AI response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}